@@ -1,150 +1,155 @@
 [build-system]
 build-backend = "hatchling.build"
-requires = ["hatchling"]
+requires = [ "hatchling" ]
 
 [project]
 name = "flowsom"
 version = "0.2.1"
 description = "The complete FlowSOM package known from R, now available in Python!"
 readme = "README.md"
-requires-python = ">=3.9"
 license = { file = "LICENSE" }
-authors = [
-    { name = "Artuur Couckuyt", email = "Artuur.Couckuyt@ugent.be" },
-    { name = "Benjamin Rombaut", email = "Benjamin.Rombaut@ugent.be" },
-    { name = "Yvan Saeys", email = "Yvan.Saeys@UGent.be" },
-    { name = "Sofie Van Gassen", email = "Sofie.VanGassen@UGent.be" },
-]
 maintainers = [
-    { name = "Artuur Couckuyt", email = "Artuur.Couckuyt@ugent.be" },
-    { name = "Benjamin Rombaut", email = "Benjamin.Rombaut@ugent.be" },
-    { name = "Yvan Saeys", email = "Yvan.Saeys@UGent.be" },
-    { name = "Sofie Van Gassen", email = "Sofie.VanGassen@UGent.be" },
+  { name = "Artuur Couckuyt", email = "Artuur.Couckuyt@ugent.be" },
+  { name = "Benjamin Rombaut", email = "Benjamin.Rombaut@ugent.be" },
+  { name = "Yvan Saeys", email = "Yvan.Saeys@UGent.be" },
+  { name = "Sofie Van Gassen", email = "Sofie.VanGassen@UGent.be" },
 ]
-urls.Documentation = "https://flowsom.readthedocs.io/en/latest/"
-urls.Source = "https://github.com/saeyslab/FlowSOM_Python"
-urls.Home-page = "https://github.com/saeyslab/FlowSOM_Python"
+authors = [
+  { name = "Artuur Couckuyt", email = "Artuur.Couckuyt@ugent.be" },
+  { name = "Benjamin Rombaut", email = "Benjamin.Rombaut@ugent.be" },
+  { name = "Yvan Saeys", email = "Yvan.Saeys@UGent.be" },
+  { name = "Sofie Van Gassen", email = "Sofie.VanGassen@UGent.be" },
+]
+requires-python = ">=3.10"
+classifiers = [
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+]
 dependencies = [
-    "mudata",
-    "numpy",
-    "matplotlib",
-    "pandas",
-    "scipy",
-    "readfcs",
-    "scikit-learn",
-    "igraph",
-    "session-info",
-    "numba",
-    "scanpy",
-    "seaborn",
-    "anndata",
-    "loguru",
+  "anndata",
+  "igraph",
+  "loguru",
+  "matplotlib",
+  "mudata",
+  "numba",
+  "numpy",
+  "pandas",
+  "readfcs",
+  "scanpy",
+  "scikit-learn",
+  "scipy",
+  "seaborn",
+  # for debug logging (referenced from the issue template)
+  "session-info2",
 ]
 
-
-[project.optional-dependencies]
-dev = ["pre-commit", "twine>=4.0.2"]
-doc = [
-    "docutils>=0.8,!=0.18.*,!=0.19.*",
-    "sphinx>=4",
-    "sphinx-book-theme>=1.0.0",
-    "myst-nb",
-    "sphinxcontrib-bibtex>=1.0.0",
-    "sphinx-autodoc-typehints",
-    "sphinxext-opengraph",
-    # For notebooks
-    "ipykernel",
-    "ipython",
-    "sphinx-copybutton",
-    # For example notebooks
-    "pytometry>=0.1.5",
+optional-dependencies.dev = [ "pre-commit", "twine>=4.0.2" ]
+optional-dependencies.doc = [
+  "docutils>=0.8,!=0.18.*,!=0.19.*",
+  # For notebooks
+  "ipykernel",
+  "ipython",
+  "myst-nb>=1.1",
+  "pandas",
+  # For example notebooks
+  "pytometry>=0.1.5",
+  # Until pybtex >0.24.0 releases: https://bitbucket.org/pybtex-devs/pybtex/issues/169/
+  "setuptools",
+  "sphinx>=4",
+  "sphinx-autodoc-typehints",
+  "sphinx-book-theme>=1",
+  "sphinx-copybutton",
+  "sphinx-tabs",
+  "sphinxcontrib-bibtex>=1",
+  "sphinxext-opengraph",
 ]
-test = ["pytest", "coverage"]
+optional-dependencies.test = [ "coverage", "pytest" ]
+urls.Documentation = "https://flowsom.readthedocs.io/en/latest/"
+urls.Home-page = "https://github.com/saeyslab/FlowSOM_Python"
+urls.Source = "https://github.com/saeyslab/FlowSOM_Python"
 
 [tool.hatch.build.targets.wheel]
-packages = ['src/flowsom']
+packages = [ 'src/flowsom' ]
 
-[tool.coverage.run]
-source = ["flowsom"]
-omit = ["**/test_*.py"]
+[tool.hatch.envs.default]
+installer = "uv"
+features = [ "dev" ]
 
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-xfail_strict = true
-addopts = [
-    "--import-mode=importlib", # allow using test files with same name
-]
+[tool.hatch.envs.docs]
+features = [ "doc" ]
+scripts.build = "sphinx-build -M html docs docs/_build {args}"
+scripts.open = "python -m webbrowser -t docs/_build/html/index.html"
+scripts.clean = "git clean -fdX -- {args:docs}"
 
-[tool.fawltydeps]
-code = ["src"]                  # Only search for imports under ./src
-deps = ["pyproject.toml"]       # Only look for declared dependencies here
-ignore_undeclared = ["flowsom"]
+[tool.hatch.envs.hatch-test]
+features = [ "test" ]
 
 [tool.ruff]
 line-length = 120
-src = ["src"]
-extend-include = ["*.ipynb"]
+src = [ "src" ]
+extend-include = [ "*.ipynb" ]
 
-[tool.ruff.format]
-docstring-code-format = true
-
-[tool.ruff.lint]
-select = [
-    "F",      # Errors detected by Pyflakes
-    "E",      # Error detected by Pycodestyle
-    "W",      # Warning detected by Pycodestyle
-    "I",      # isort
-    "D",      # pydocstyle
-    "B",      # flake8-bugbear
-    "TID",    # flake8-tidy-imports
-    "C4",     # flake8-comprehensions
-    "BLE",    # flake8-blind-except
-    "UP",     # pyupgrade
-    "RUF100", # Report unused noqa directives
+format.docstring-code-format = true
+lint.select = [
+  "B",      # flake8-bugbear
+  "BLE",    # flake8-blind-except
+  "C4",     # flake8-comprehensions
+  "D",      # pydocstyle
+  "E",      # Error detected by Pycodestyle
+  "F",      # Errors detected by Pyflakes
+  "I",      # isort
+  "RUF100", # Report unused noqa directives
+  "TID",    # flake8-tidy-imports
+  "UP",     # pyupgrade
+  "W",      # Warning detected by Pycodestyle
 ]
-ignore = [
-    # line too long -> we accept long comment lines; formatter gets rid of long code lines
-    "E501",
-    # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
-    "E731",
-    # allow I, O, l as variable names -> I is the identity matrix
-    "E741",
-    # Missing docstring in public package
-    "D104",
-    # Missing docstring in public module
-    "D100",
-    # Missing docstring in __init__
-    "D107",
-    # Errors from function calls in argument defaults. These are fine when the result is immutable.
-    "B008",
-    # __magic__ methods are are often self-explanatory, allow missing docstrings
-    "D105",
-    # first line should end with a period [Bug: doesn't work with single-line docstrings]    "D400",
-    # First line should be in imperative mood; try rephrasing
-    "D401",
-    ## Disable one in each pair of mutually incompatible rules
-    # We don’t want a blank line before a class docstring
-    "D203",
-    # We want docstrings to start immediately after the opening triple quote
-    "D213",
+lint.ignore = [
+  "B008", # Errors from function calls in argument defaults. These are fine when the result is immutable.
+  "D100", # Missing docstring in public module
+  "D104", # Missing docstring in public package
+  "D105", # __magic__ methods are often self-explanatory, allow missing docstrings
+  "D107", # Missing docstring in __init__
+  # Disable one in each pair of mutually incompatible rules
+  "D203", # We don’t want a blank line before a class docstring
+  "D213", # <> We want docstrings to start immediately after the opening triple quote
+  "D400", # first line should end with a period [Bug: doesn’t work with single-line docstrings]
+  "D401", # First line should be in imperative mood; try rephrasing
+  "E501", # line too long -> we accept long comment lines; formatter gets rid of long code lines
+  "E731", # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
+  "E741", # allow I, O, l as variable names -> I is the identity matrix
+]
+lint.per-file-ignores."*/__init__.py" = [ "F401" ]
+lint.per-file-ignores."docs/*" = [ "I" ]
+lint.per-file-ignores."tests/*" = [ "D" ]
+lint.pydocstyle.convention = "numpy"
+
+[tool.pytest.ini_options]
+testpaths = [ "tests" ]
+xfail_strict = true
+addopts = [
+  "--import-mode=importlib", # allow using test files with same name
 ]
 
-[tool.ruff.lint.pydocstyle]
-convention = "numpy"
+[tool.coverage.run]
+source = [ "flowsom" ]
+omit = [ "**/test_*.py" ]
 
-[tool.ruff.lint.per-file-ignores]
-"docs/*" = ["I"]
-"tests/*" = ["D"]
-"*/__init__.py" = ["F401"]
+[tool.fawltydeps]
+code = [ "src" ]                  # Only search for imports under ./src
+deps = [ "pyproject.toml" ]       # Only look for declared dependencies here
+ignore_undeclared = [ "flowsom" ]
 
 [tool.cruft]
 skip = [
-    "tests",
-    "src/**/__init__.py",
-    "src/**/basic.py",
-    "docs/api.md",
-    "docs/changelog.md",
-    "docs/references.bib",
-    "docs/references.md",
-    "docs/notebooks/example.ipynb",
+  "tests",
+  "src/**/__init__.py",
+  "src/**/basic.py",
+  "docs/api.md",
+  "docs/changelog.md",
+  "docs/references.bib",
+  "docs/references.md",
+  "docs/notebooks/example.ipynb",
 ]
